+++ app/code/ThirdParty/BlogArticle/Controller/Adminhtml/Post/NewAction.php
@@ -1,0 +1,33 @@
+<?php
+namespace ThirdParty\BlogArticle\Controller\Adminhtml\Post;
+
+use Magento\Backend\App\Action;
+use Magento\Backend\App\Action\Context;
+use Magento\Backend\Model\View\Result\ForwardFactory;
+
+class NewAction extends Action
+{
+    const ADMIN_RESOURCE = 'ThirdParty_BlogArticle::posts';
+
+    /**
+     * @var ForwardFactory
+     */
+    private $resultForwardFactory;
+
+    public function __construct(
+        Context $context,
+        ForwardFactory $resultForwardFactory
+    ) {
+        parent::__construct($context);
+        $this->resultForwardFactory = $resultForwardFactory;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function execute()
+    {
+        $resultForward = $this->resultForwardFactory->create();
+        return $resultForward->forward('edit');
+    }
+}