+++ app/code/ThirdParty/BlogArticle/Model/PostFactory.php
@@ -1,0 +1,39 @@
+<?php
+namespace ThirdParty\BlogArticle\Model;
+
+use Magento\Framework\ObjectManagerInterface;
+
+class PostFactory
+{
+    /**
+     * @var ObjectManagerInterface
+     */
+    private $objectManager;
+
+    /**
+     * @var string
+     */
+    private $instanceName;
+
+    public function __construct(
+        ObjectManagerInterface $objectManager,
+        $instanceName = Post::class
+    ) {
+        $this->objectManager = $objectManager;
+        $this->instanceName = $instanceName;
+    }
+
+    /**
+     * @param array $data
+     * @return Post
+     */
+    public function create(array $data = [])
+    {
+        /** @var Post $post */
+        $post = $this->objectManager->create($this->instanceName);
+        if (!empty($data)) {
+            $post->setData($data);
+        }
+        return $post;
+    }
+}